Should have cast to 'long' rather than 'unsigned int'.

+++ coders/xcf.c
@@ -1629,7 +1629,7 @@ static Image *ReadXCFImage(const ImageInfo *image_info,ExceptionInfo *exception)
 	    */
 	    image->next=layer_info[number_layers-2].image;
 	    layer_info[number_layers-2].image->previous=image;
-	    for (j=(signed int) number_layers-2; j >= 0; j--)
+	    for (j=(long) number_layers-2; j >= 0; j--)
 	      {
 		if (j > 0)
 		  layer_info[j].image->next=layer_info[j-1].image;
